// src/components/FormBuilder/types/all-fields.d.ts
type BaseField = {
  zod?: z.ZodType;
  fieldsToWatch?: any[];
};

type ComponentPropsFunction<T, P> = ({
  WatchValues,
}: {
  WatchValues: T | Record<string, any>;
} & UseFormReturn<T | any>) => Omit<P, "control" | "name">;

type FieldConfig<T, P> = BaseField & {
  componentProps: Omit<P, "control" | "name"> | ComponentPropsFunction<T, P>;
};

type FieldArrayFunctionProps<T> = {
  WatchValues: Record<string, any>;
  fieldArrayHooks: UseFieldArrayReturn<T, string, "id">;
  FormHooks: UseFormReturn<T>;
  index: number;
  id: string;
};
// ...
type Field<T> =
  | (FieldConfig<T, InputFieldProps> & {
      type: "input";
    })
  | (FieldConfig<T, PhoneNumberInputFieldProps> & {
      type: "phone-number";
    })
  | (FieldConfig<T, CheckboxFieldProps> & {
      type: "checkbox";
    })
  | (FieldConfig<T, SwitchFieldProps> & {
      type: "switch";
    })
  | (FieldConfig<T, DateFieldProps> & {
      type: "datetime-local";
    })
  | (FieldConfig<T, RadioFieldProps> & {
      type: "radio";
    })
  | (FieldConfig<T, TextEditorFieldProps> & {
    type: "text-editor";
  })
  | (FieldConfig<T, FileUploadFieldProps> & {
      type: "file-upload";
    })
  | (FieldConfig<T, AutocompleteFieldProps> & {
      type: "autocomplete";
    })
  | (FieldConfig<T, AutocompleteFieldServerProps> & {
      type: "autocomplete-server";
    })
  | (FieldConfig<T, AutocompleteFieldOneTimeFetchProps> & {
      type: "autocomplete-one-time-fetch";
    })
  | (FieldConfig<T, Parameters<K>[0]> & {
      type: "custom";
      component: K;
    })
  | (BaseField & {
      type: "field-array";
      containerClassName?: string;
      errorClassName?: string;
      fields: (props: FieldArrayFunctionProps<T>) => Record<string, Field<T>>;
    });
// ...
type AllFieldsTypes = Field["type"];